Extracted console interaction into separate library for examples.
- To be reused in Azure block.
