+++ chat-example.sln
@@ -33,6 +33,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "chat-server", "examples\azu
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "chat-client", "examples\azure-websockets\chat-client\chat-client.csproj", "{5B7D1F74-56F5-4EBE-BADE-79A01152452F}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "console-interaction", "examples\console-interaction\console-interaction.csproj", "{70389050-FB40-41F4-9CE0-485391089223}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -166,6 +168,18 @@ Global
 		{5B7D1F74-56F5-4EBE-BADE-79A01152452F}.Release|x64.Build.0 = Release|Any CPU
 		{5B7D1F74-56F5-4EBE-BADE-79A01152452F}.Release|x86.ActiveCfg = Release|Any CPU
 		{5B7D1F74-56F5-4EBE-BADE-79A01152452F}.Release|x86.Build.0 = Release|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Debug|x64.Build.0 = Debug|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Debug|x86.Build.0 = Debug|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Release|x64.ActiveCfg = Release|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Release|x64.Build.0 = Release|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Release|x86.ActiveCfg = Release|Any CPU
+		{70389050-FB40-41F4-9CE0-485391089223}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{90422F2B-B999-40CC-BCC3-BBE9161B3595} = {F5B3CF68-9413-4CA3-AC9D-43831F02C678}
@@ -180,5 +194,6 @@ Global
 		{72F45459-17DE-466D-9C8D-750131A3F044} = {B75E9D57-A885-4791-A09F-F51A6125C38D}
 		{C86DB07E-8EC5-4ACB-83F4-E752FF4E258F} = {72F45459-17DE-466D-9C8D-750131A3F044}
 		{5B7D1F74-56F5-4EBE-BADE-79A01152452F} = {72F45459-17DE-466D-9C8D-750131A3F044}
+		{70389050-FB40-41F4-9CE0-485391089223} = {B75E9D57-A885-4791-A09F-F51A6125C38D}
 	EndGlobalSection
 EndGlobal
