+++ examples/azure-websockets/chat-client/chat-client.csproj
@@ -12,6 +12,7 @@
     <ProjectReference Include="..\..\..\src\protocol\protocol.csproj" />
     <ProjectReference Include="..\..\..\src\client-library\client-library.csproj" />
     <ProjectReference Include="..\..\protobuf-protocol\protobuf-protocol.csproj" />
+    <ProjectReference Include="..\..\console-interaction\console-interaction.csproj" />
   </ItemGroup>
 
   <ItemGroup>
