+++ examples/console-interaction/Commands/ICommand.cs
@@ -1,0 +1,6 @@
+namespace Lemvik.Example.Chat.Client.Examples.Commands
+{
+    public interface ICommand
+    {
+    }
+}
