+++ examples/console-interaction/Commands/ICommandsSource.cs
@@ -1,7 +1,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Lemvik.Example.Chat.Client.Example.TCP.Commands
+namespace Lemvik.Example.Chat.Client.Examples.Commands
 {
     public interface ICommandsSource
     {
