+++ examples/console-interaction/Commands/InfoCommand.cs
@@ -1,0 +1,6 @@
+namespace Lemvik.Example.Chat.Client.Examples.Commands
+{
+    public class InfoCommand : ICommand
+    {
+    }
+}
