+++ examples/console-interaction/Commands/JoinRoomCommand.cs
@@ -1,4 +1,4 @@
-namespace Lemvik.Example.Chat.Client.Example.TCP.Commands
+namespace Lemvik.Example.Chat.Client.Examples.Commands
 {
     public class JoinRoomCommand : ICommand
     {
