+++ examples/console-interaction/IConsoleClient.cs
@@ -1,0 +1,10 @@
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Lemvik.Example.Chat.Client.Examples
+{
+    public interface IConsoleClient
+    {
+        Task InteractAsync(IChatClient client, CancellationToken token = default);
+    }
+}
