--- examples/simple-tcp/chat-client/Commands/ICommand.cs
@@ -1,6 +1,0 @@
-namespace Lemvik.Example.Chat.Client.Example.TCP.Commands
-{
-    public interface ICommand
-    {
-    }
-}
