+++ examples/simple-tcp/chat-client/chat-client.csproj
@@ -4,6 +4,7 @@
     <ProjectReference Include="..\..\..\src\protocol\protocol.csproj" />
     <ProjectReference Include="..\..\..\src\client-library\client-library.csproj" />
     <ProjectReference Include="..\..\protobuf-protocol\protobuf-protocol.csproj" />
+    <ProjectReference Include="..\..\console-interaction\console-interaction.csproj" />
   </ItemGroup>
 
 
@@ -18,7 +19,7 @@
   </ItemGroup>
     
   <ItemGroup>
-    <None Include="appsettings.json" CopyToPublishDirectory="Always" CopyToOutputDirectory="Always"/>
+    <None Include="appsettings.json" CopyToPublishDirectory="Always" CopyToOutputDirectory="Always" />
   </ItemGroup>
 
   <PropertyGroup>
